#3398 mv tests to own file
